Move fetch results to the server directory
